#96 move ProjectShortInfo to format package
